pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "utility for comparing changes in json and yaml files"
authors = ["Ihar Kostyuchik <kostyuchik22@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest-cov = "^4.0.0"
pyyaml = "^6.0.1"
pytest = "^7.4.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"